+++ src/main/java/com/healingpill/controller/MemberController.java
@@ -25,9 +25,9 @@ public class MemberController {
         logger.debug("로그인 페이지 이동");
     }
 
-    // 로그인 페이지 이동
-    @RequestMapping(value = "/ex12JoinSuccess", method = RequestMethod.POST)
-    public void ex12JoinSuccessPOST() {
+    // 로그인 성공 페이지 이동
+    @RequestMapping(value = "/memberJoinSuccess", method = RequestMethod.POST)
+    public void memberJoinSuccessPOST() {
 
         logger.debug("로그인 성공 페이지 이동");
     }
