+++ src/main/java/com/healingpill/member/MemberDao.java
@@ -1,4 +1,4 @@
-package com.healingpill.domain.member;
+package com.healingpill.member;
 
 import com.healingpill.config.DB;
