+++ src/main/webapp/WEB-INF/dispatcher-servlet.xml
@@ -18,10 +18,12 @@
 
     <context:component-scan base-package="com.healingpill.config"/>
     <context:component-scan base-package="com.healingpill.controller"/>
-    <context:component-scan base-package="com.healingpill.service"/>
-    <context:component-scan base-package="com.healingpill.util"/>
-    <context:component-scan base-package="com.healingpill.model"/>
     <context:component-scan base-package="com.healingpill.mapper"/>
+    <context:component-scan base-package="com.healingpill.member"/>
+    <context:component-scan base-package="com.healingpill.model"/>
+    <context:component-scan base-package="com.healingpill.service"/>
+    <context:component-scan base-package="com.healingpill.servlets"/>
+
 
     <bean class = "org.springframework.web.servlet.view.InternalResourceViewResolver">
         <property name="prefix" value="/WEB-INF/views/"></property>
